Update workflowr to 1.0.1 with wflow_update().
Some of the session information chunks got left behind because they were
slightly modified (e.g. the name of the chunk was changed). This will
result in duplicate session information chunks, so should be removed
when making other updates to the Rmd (there are other options, but this
is the simplest).